update to render function (for programmatic positioning of multi-slider)
when the first slider moves first and intersects with the second slider
it stops moving.  As a results, if the first slider needs to move past
the second slider we need to make sure the first slider gets to the
proper position *after* the second slider has moved to its position.